Refactor métodos en Track y SpotifyCsvParser para mejorar la conversión de datos a conjuntos de Weka. Cambiar el tipo de salida en ArtistsPopularities a FloatWritable y agregar métodos de entrenamiento y predicción sobre instancias de Weka en ClassificationModel.

+++ src/main/java/mapreduce/ArtistsPopularities.java
@@ -78,7 +78,7 @@ public class ArtistsPopularities extends Configured implements Tool {
         conf.setJobName("Get the relationship between artists popularity and their songs popularity");
 
         conf.setOutputKeyClass(Text.class);
-        conf.setOutputValueClass(IntWritable.class);
+        conf.setOutputValueClass(FloatWritable.class);
 
         conf.setMapperClass(ArtistsPopularitiesMapper.class);
         conf.setReducerClass(ArtistsPopularitiesReducer.class);
